Add package & testbench for simple dictionary protected type
A new packae DictP.vhd is added which contains a simple implementation
of a dictionary type. It consists of a protected type t_dict which
contains procedures & functions which serve as methods to interact with
the protected type. The t_dict type is realised by a (double)
linked-list using access types.
The key type is string of arbitrary size, the data type is
std_logic_vector of arbitrary size. The t_dict type provides following
interface methods:

  * set: set a key/value pair
  * get: get a key/value pair
  * del: remove a key/value pair
  * init: initialise the dictionary
  * clear: remove all key/value pairs
  * hasKey: check if a key exists in the dictionary
  * size: get number of key/value pairs